+++ R/utils.R
@@ -1,0 +1,5 @@
+# Specific Version of Stop for roperators
+rop_stop <- function(...){
+  stop("roperators: \n ", ...,
+       call. = FALSE)
+}