+++ NEWS.md
@@ -1,8 +1,24 @@
-# roperators 1.1.1
+# roperators 1.2.0
 
 * Cleaned up roxygen comments, DESCRIPTION, and NAMESPACE files.
 * Added a `NEWS.md` file to track changes to the package.
+* New complete cases functions (`length_cc()`, `min_cc()`, `max_cc()`,
+  `range_cc()`, `all_cc()`, `any_cc()`, `sum_cc()`, `prod_cc()`, `mean_cc()`,
+  `median_cc()`, `var_cc()`, `cov_cc()`, `cor_cc()`, `sd_cc()`,
+  `weighted_mean_cc()`, `quantile_cc()`, `IQR_cc()`, `mad_cc()`,
+  `rowSums_cc()`, `colSums_cc()`, `rowMeans_cc()`, `colMeans_cc()`) that set
+  default argument to standard base/stats functions for only calculating on
+  complete cases.
 * New `get_os()`, `is.os_mac`, `is.os_win`, `is.os_lnx`, and `is.os_unx` helper
   functions to quickly determine operating system.
-* New `n_unique()` function to count unique items.
-* New `get_most_frequent()` function to get most frequent thing(s) in x
+* New `is_txt_file()`, `is_csv_file()`, `is_excel_file()`, `is_r_file()`, 
+  `is_rdata_file()`, `is_rda_file()`, `is_spss_file()`, and `check_ext_against()`
+  functions to quickly determine if file has specific file extensions.
+* New `is.R_Revo()` and `is.RStudio()` functions to determine whether Revolution
+  R and RStudio are installed.
+* New `n_unique()` function to count unique items and `n_unique_cc()` version
+  for complete cases.
+* New `is.constant()` and `is.binary()` functions to indicate whether a vector
+  is constant (contains only ONE value type, ignoring NA) or binary (contains
+  at most TWO value types, ignoring NA).
+* New `get_most_frequent()` function to get most frequent thing(s) in x.